refactor to support small field executor and bootloader
adds a BootloaderImpl trait that should be implemented by the small/large field
variants of the bootloader
